requirements.txt
python-telegram-bot==20.3
selenium==4.8.3
pandas==1.5.3
openpyxl==3.1.2